test: verify event subprocess subscription migration and re-subscription
So far we handled migrating active event subprocesses. But there is also a case where users
should be able to migrate event subprocesses that are not active yet. In that case, if the
start event of the event subprocess is a message start event, a message subscription is created
for the enclosing process.

There are 3 cases that we need to support while migrating message subscriptions of event
subprocesses:

1) Migrating the message subscription of the event subprocess
In this case, a mapping instruction for message start events should be provided. Therefore, we will
migrate the subscription from source to target.

2) Subscribing to a message subscription of the event subprocess in the target
When the target process contains an event subprocess with a message start event and if no mapping
defined for it, we need to subscribe to the message start event of the target process.

3) Unsubscribing from a message subscription of event subprocess in the source
When the source process contains an event subprocess with a message start event and if no mapping
defined for it, we need to unsubscribe from the message start event of the source process.

All cases above are already supported through `ProcessInstanceMigrationMigrateProcessor
.handleCatchEvents()` method. There, we handle the case if the message subscription is migrated, or
we need subscribing/unsubscribing.

Lastly, please note that when there are message event subprocesses in both source and the target
process with a same message name, a migration mapping has to be defined for these start events.
Because, we have a known limitation that we cannot resubscribe to the same message name during the
migration. Until the limitation is resolved, a workaround is to provide a mapping for the catch
events that have the same message name.