fix: do not create subscription for start event of interrupting process
The fix to prevent the situation mentioned in previous commit:

With the current structure of handling catch events in `ProcessInstanceMigrationMigrateProcessor`,
when there is an unmapped catch event in the target, we subscribe to that event. However, it is a
wrong behaviour for interrupting event subprocesses. While migrating interrupting event
subprocesses, a new subscription for the start event shouldn't be created because interrupting
subprocess flow should be taken only once.

When an interrupting subprocess is entered, all elements of the enclosing process except the
interrupting subprocess is terminated. But when `getEvents()` method on the element is called, the
start event of the contained event process is still returned. This commit ensures that we do not
re-create a subscription for that start event.