feat: migrate timer event subprocess
To support timer event subprocess migration, it is enough to add TIMER to supported event
subprocess types. It is because we have a generic implementation in handle catch events where we
migrated mapped timer events, subscribe to the ones that are not mapped and exists in target, and
unsubscribe from the ones that are not mapped and exist in the source.