test: ensure migration can resolve xor incident
A process instance may be waiting at an Exclusive Gateway (XOR) due to
an Incident created as a result of a faulty condition expression. This
modeling mistake can be resolved by the user, by deploying the improved
version and migrating the process instance over to it. When resolving
the incident, the exclusive gateway should evaluate the conditions as
defined by the target process definition.