refactor: simplify logic
The double usage of element.getOutgoing seemed strange when I was first
reading it, and I wondered whether it perhaps was retrieving some
different value. But, it was just the first entry from the list.

This code can be simplified by reusing the local variable
outgoingSequenceFlows, and storing its first entry in another local
variable for reuse. Reading the code line-for-line is now easier to
follow.