fix: reject migrate message subscription command
In a distributed (partitioned) system, we don't know whether the message
subscription still exists on the partition when the command to migrate
it is received and being processed.

As the context may change in our system, command processors must always
consider the current state to decide how to process the command.

For migrate message subscription commands, the processor must check
whether the subscription exists in the state, and then mark it as
migrated. It's also important to not blindly take all data from the
command. It may contain outdated information. Instead, we should use the
information from the state about the subscription for the MIGRATED
event, and adjust only the relevant data: bpmn process id +
interrupting.

Note that the subscription is not found by key but by element instanced
key and message name.

If the subscription cannot be found at all, we must reject the command.
In this case, we need to consider whether we want to acknowledge the
command (if it was distributed from another partition). At this time, we
don't have a reject distributed command operation yet, so we can either
acknowledge it or not. If we acknowledge it, then the migrate command
will not be retried, otherwise it will. If the message subscription
cannot be found it must mean that the message subscription was already
correlated and removed from the state. As far as I know there are no
other cases at this time. There is no need to retry, so we can
acknowledge the command.