feat: transparent fallback to static routing info
Until the routing info is updated for the first time, we can transparently fall back to static configuration.

This approach is easier than adding a database migration or initializing somewhere else (e.g. Engine#init) because it
does not require any API changes.
It also ensures that the engine state is not modified as long as the partition scaling feature is disabled. This makes
backwards compatibility much easier.